requirements.txt
# === Core dependencies ===
numpy
pandas
scikit-learn
umap-learn

# === Web server ===
Flask
Flask-Cors
Flask-SocketIO
gunicorn
python-socketio            

# === Data validation ===
pydantic>=2.0
typing-extensions

# === HTTP + env ===
requests
python-dotenv
packaging

# === Embedding + local LLM ===
sentence-transformers
transformers
torch
accelerate          
safetensors        
huggingface-hub     
rank_bm25

# === Vector DB (Zilliz Cloud / Milvus) ===
pymilvus

# === LangChain ecosystem ===
langchain
langchain-core
langchain-community
langchain-openai   

# === Optional utilities ===
tqdm
colorama

# === Logging ===
google-cloud-logging